requirements.txt
aiogram==3.*
python-dotenv==1.0.1
aiohttp==3.11.11